+++ CHANGELOG.md
@@ -8,6 +8,12 @@ with the `YY.MINOR.MICRO` scheme.
 
 <img src="https://img.shields.io/badge/calver-YY.MINOR.MICRO-22bfda.svg">
 
+## [25.3.8] - Unreleased
+
+- Fixed an out-of-memory crash when using Response::encode_rmp with large
+  collections in plugin API (see also https://git.picodata.io/core/picodata/-/issues/2028).
+
+
 ## [25.3.7] - 2025-10-16
 
 ### Fixes